+++ pkg/cookiebanner/pattern_analysis_worker.go
