+++ pkg/cookiebanner/pattern_analysis_worker_test.go
