+++ pkg/cookiebanner/tracker_mapping_worker.go
@@ -12,7 +12,7 @@
 // OTHER TORTIOUS ACTION, ARISING OUT OF OR IN CONNECTION WITH THE USE OR
 // PERFORMANCE OF THIS SOFTWARE.
 
-package probo
+package cookiebanner
 
 import (
 	"context"
